requirements.txt
feedparser~=6.0.11
PyYAML~=6.0.1
lxml~=5.2.0
beautifulsoup4~=4.12.3
python-dateutil~=2.9.0
mysql-connector-python~=8.3.0
requests~=2.31.0